+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * No longer test the `setup-dart` action on pre-2.12 SDKs.
 * Upgrade JS interop code to use extension types
   (the new name for inline classes).
+* The upcoming rename of the `be` channel to `main` is now supported with
+  forward compatibility that switches when the rename happens.
 
 ## v1.5.0