+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v1.5.1-dev
+## v1.5.1
 
 * No longer test the `setup-dart` action on pre-2.12 SDKs.
 * Upgrade JS interop code to use extension types